Isolate spy recording behind spyrecord build tag
Move spy infrastructure out of production code using build-tagged file
pairs. Each wrapper package (git, gh, claude) now has record_noop.go
(empty no-op, compiled by default) and record_spy.go (real recording,
compiled only with -tags spyrecord). Production binary no longer
imports the spy package or runs init() checks.
