+++ internal/gh/gh.go
@@ -4,21 +4,9 @@ package gh
 import (
 	"encoding/json"
 	"fmt"
-	"os"
 	"os/exec"
 	"strings"
-
-	"github.com/pavelpascari/sdf/internal/spy"
 )
-
-var fullSpy *spy.Recorder
-
-func init() {
-	if dir := os.Getenv("SDF_SPY_DIR"); dir != "" {
-		Spy = spy.NewRecorderFor(dir, "sdf", "gh")
-		fullSpy = spy.NewRecorder(dir, "full")
-	}
-}
 
 // PRInfo represents pull request information from gh.
 type PRInfo struct {
@@ -35,24 +23,17 @@ type PRInfo struct {
 // Tests can override this to point at a fake binary.
 var Binary = "gh"
 
-// Spy, when non-nil, records every invocation for later analysis.
-// Enable during E2E tests to capture real API responses.
-var Spy *spy.Recorder
-
 // run executes a gh command and returns its trimmed stdout.
 func run(args ...string) (string, error) {
 	cmd := exec.Command(Binary, args...)
 	out, err := cmd.CombinedOutput()
 	output := strings.TrimSpace(string(out))
 
-	if Spy != nil {
-		exitCode := 0
-		if err != nil {
-			exitCode = 1
-		}
-		Spy.Record(args, output, exitCode)
-		fullSpy.RecordAs("sdf", "gh", args, output, exitCode)
+	exitCode := 0
+	if err != nil {
+		exitCode = 1
 	}
+	recordRun(args, output, exitCode)
 
 	if err != nil {
 		return output, fmt.Errorf("gh %s: %s", strings.Join(args, " "), output)
