+++ internal/testutil/fakebin.go
@@ -8,8 +8,6 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
-
-	"github.com/pavelpascari/sdf/internal/spy"
 )
 
 // FakeBin creates a shell script at dir/<name> that logs its arguments
@@ -118,17 +116,3 @@ func SetBinary(t *testing.T, target *string, fakePath string) {
 	t.Cleanup(func() { *target = orig })
 }
 
-// SetSpy sets a package-level Spy variable and registers cleanup.
-// Usage:
-//
-//	rec := spy.NewRecorder(dir, "gh")
-//	SetSpy(t, &gh.Spy, rec)
-func SetSpy(t *testing.T, target **spy.Recorder, rec *spy.Recorder) {
-	t.Helper()
-	orig := *target
-	*target = rec
-	t.Cleanup(func() {
-		rec.Close()
-		*target = orig
-	})
-}
